+++ typescript/src/index.ts
@@ -1,5 +1,8 @@
 /* tslint:disable */
 /* eslint-disable */
+// Hand-maintained (see .openapi-generator-ignore): the generator's barrel plus
+// ./field, which supplies the tenant headers the contract does not describe.
 export * from './runtime';
 export * from './apis/index';
 export * from './models/index';
+export * from './field';